bug for int total

+++ invoice/post_add.php
@@ -540,8 +540,8 @@ $MSTOTAL=floatval($MSericeInPrice)+floatval($MSericeOutPrice)+floatval($MSericeB
 #########################################################################
 
 $TOTAL=$MSTOTAL+$SSTOTAL ;
-$VAT=intval(($TOTAL*$ShipVAT)*($vatP));
-$VAT_TOTAL=intval($VAT+$TOTAL);
+$VAT=floatval(($TOTAL*$ShipVAT)*($vatP));
+$VAT_TOTAL=floatval($VAT+$TOTAL);
 
     if($debug){echo "<b>M. TOTAL :</b>".$MSTOTAL."<br>";}
     if($debug){echo "<b>S. TOTAL :</b>".$SSTOTAL."<br>";}
